# docs/api/external-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sasai Wallet \u2014 Partner API",
    "version": "0.1.0",
    "description": "Partner-facing external API. Authenticate with X-Sasai-Api-Key plus an HMAC X-Sasai-Signature over the raw request body."
  },
  "paths": {
    "/api/v1/external/users": {
      "post": {
        "tags": [
          "external"
        ],
        "summary": "Create a user",
        "description": "Create a user in the API key's tenant, reusing the identity service.\n\nThe tenant comes from `principal` (the key), never the body. On a retry\nthat collides with an existing identifier, returns the existing user (200)\nrather than 409 \u2014 the identifier is the natural idempotency key.",
        "operationId": "create_external_user_api_v1_external_users_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-Sasai-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sasai-Api-Key"
            }
          },
          {
            "name": "X-Sasai-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 2048
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sasai-Signature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalCreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "200": {
            "description": "Idempotent replay \u2014 the user already existed; returned unchanged."
          },
          "401": {
            "description": "Missing or invalid API key / signature."
          },
          "422": {
            "description": "Validation error \u2014 e.g. no email/phone identifier, or a missing Idempotency-Key header."
          },
          "429": {
            "description": "Per-key rate limit exceeded."
          }
        }
      }
    },
    "/api/v1/external/fund": {
      "post": {
        "tags": [
          "external"
        ],
        "summary": "Fund a user's wallet",
        "description": "Credit a user's wallet in the API key's tenant.\n\nTenant comes from the key, never the body. The `Idempotency-Key` header is\nrequired and is used as the ledger key \u2014 a retry returns the original result\nwithout double-crediting.",
        "operationId": "fund_external_api_v1_external_fund_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-Sasai-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sasai-Api-Key"
            }
          },
          {
            "name": "X-Sasai-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 2048
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sasai-Signature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalFundRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key / signature."
          },
          "404": {
            "description": "The identifier does not resolve to a user/wallet in the key's tenant."
          },
          "409": {
            "description": "Insufficient funds, or nothing to withdraw (empty wallet)."
          },
          "422": {
            "description": "Validation error, or a configured limit was exceeded."
          },
          "429": {
            "description": "Per-key rate limit exceeded."
          }
        }
      }
    },
    "/api/v1/external/withdraw": {
      "post": {
        "tags": [
          "external"
        ],
        "summary": "Withdraw from a user's wallet",
        "description": "Debit a user's wallet in the API key's tenant.\n\nSend `amount`, or `withdraw_all: true` (no amount) to pull the full\navailable balance. Tenant from the key; idempotent on the required\n`Idempotency-Key`.",
        "operationId": "withdraw_external_api_v1_external_withdraw_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-Sasai-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sasai-Api-Key"
            }
          },
          {
            "name": "X-Sasai-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 2048
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sasai-Signature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalWithdrawRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawFromUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key / signature."
          },
          "404": {
            "description": "The identifier does not resolve to a user/wallet in the key's tenant."
          },
          "409": {
            "description": "Insufficient funds, or nothing to withdraw (empty wallet)."
          },
          "422": {
            "description": "Validation error, or a configured limit was exceeded."
          },
          "429": {
            "description": "Per-key rate limit exceeded."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ExternalCreateUserRequest": {
        "properties": {
          "identifiers": {
            "items": {
              "$ref": "#/components/schemas/ExternalIdentifierIn"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Identifiers"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserProfileIn"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "identifiers"
        ],
        "title": "ExternalCreateUserRequest",
        "description": "Partner-facing create-user payload.\n\nNo `tenant_id` (derived from the API key), and no `user_type` /\n`parent_user_id` \u2014 the endpoint forces `consumer` with no parent so a\npartner can't self-assign a limit/pricing tier or graft the tenant\nhierarchy (S7 H1)."
      },
      "ExternalFundRequest": {
        "properties": {
          "identifier_type": {
            "type": "string",
            "enum": [
              "phone",
              "email",
              "account_number",
              "card_number"
            ],
            "title": "Identifier Type"
          },
          "identifier_value": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Identifier Value"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2,
            "title": "Currency"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "identifier_type",
          "identifier_value",
          "amount",
          "currency"
        ],
        "title": "ExternalFundRequest",
        "description": "Partner-facing fund payload \u2014 credits an existing user's wallet.\n\nNo `tenant_id` (derived from the API key). `extra='forbid'` rejects any\nunexpected field outright (BOPLA hardening, mirrors Epic 17 S7). The target\nis resolved by identifier and is always a user's financial_wallet."
      },
      "ExternalIdentifierIn": {
        "properties": {
          "identifier_type": {
            "type": "string",
            "enum": [
              "phone",
              "email",
              "account_number",
              "card_number"
            ],
            "title": "Identifier Type"
          },
          "identifier_value": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Identifier Value"
          }
        },
        "type": "object",
        "required": [
          "identifier_type",
          "identifier_value"
        ],
        "title": "ExternalIdentifierIn",
        "description": "A partner-supplied identifier \u2014 no `verified` flag.\n\nPartners cannot assert that a phone/email is verified; the platform only\nmarks an identifier verified through its own OTP flow."
      },
      "ExternalWithdrawRequest": {
        "properties": {
          "identifier_type": {
            "type": "string",
            "enum": [
              "phone",
              "email",
              "account_number",
              "card_number"
            ],
            "title": "Identifier Type"
          },
          "identifier_value": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Identifier Value"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "withdraw_all": {
            "type": "boolean",
            "title": "Withdraw All",
            "default": false
          },
          "currency": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2,
            "title": "Currency"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "identifier_type",
          "identifier_value",
          "currency"
        ],
        "title": "ExternalWithdrawRequest",
        "description": "Partner-facing withdraw payload \u2014 debits a user's wallet.\n\nSupply exactly one of `amount` or `withdraw_all`; `withdraw_all=true` (with\nno amount) pulls the full available balance. No `tenant_id` (from the key).\nOnly ever targets the user's financial_wallet, never a system wallet."
      },
...
      "FundUserResponse": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transaction Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "new_balance": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "New Balance"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "user_id",
          "amount",
          "currency",
          "new_balance"
        ],
        "title": "FundUserResponse",
        "description": "Result of an admin top-up."
      },
...
      "IdentifierOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "identifier_type": {
            "type": "string",
            "title": "Identifier Type"
          },
          "identifier_value": {
            "type": "string",
            "title": "Identifier Value"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          }
        },
        "type": "object",
        "required": [
          "id",
          "identifier_type",
          "identifier_value",
          "verified"
        ],
        "title": "IdentifierOut",
        "description": "An identifier echoed back on responses."
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "user_type": {
            "type": "string",
            "title": "User Type"
          },
          "parent_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent User Id"
          },
          "identifiers": {
            "items": {
              "$ref": "#/components/schemas/IdentifierOut"
            },
            "type": "array",
            "title": "Identifiers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "status",
          "user_type",
          "identifiers"
        ],
        "title": "UserOut",
        "description": "User resource returned from the API."
      },
      "UserProfileIn": {
        "properties": {
          "first_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          }
        },
        "type": "object",
        "title": "UserProfileIn",
        "description": "Optional profile data provided at registration."
      },
...
      "WithdrawFromUserResponse": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transaction Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "new_balance": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "New Balance"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "user_id",
          "amount",
          "currency",
          "new_balance"
        ],
        "title": "WithdrawFromUserResponse",
        "description": "Result of an admin withdraw."
      }
    }
  }
}